pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "origin-policy"
version = "1.0.0"
description = "Policy Gate v1 — Tier 0 absolute rules validator (R1-R5) for AI output gating"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Proprietary" }
authors = [{ name = "originaiagent" }]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
policy-gate = "origin_policy.policy_gate:main"

[tool.setuptools]
packages = ["origin_policy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
